Add sorting for Drawable z order
Windows and sprites have a z prop that defines the z position it will
be rendered in order to define which content will be above of the
others.

This adds a method that sorts the drawable vector before render when
zDirty flag is true, and a method to mark that z order has change.